Sync the fork with upstream releases v1.2.2 + v1.3.0. Resolutions:

- workspace members: keep gui-app-decred/decred-core alongside new
  passport-drive
- launcher card: adopt BitcoinPriceDetails, keep the Decred mock
  display and green sparkline (color-parameterized draw_graph_rgb
  re-added over the new stripe/scrub renderer)
- Decred QR ingest: ScanQrResult moved to struct variants
- Cargo.lock: v1.3.0 base plus the Decred crates

+++ api/i2c/src/messages.rs
@@ -23,5 +23,5 @@ impl FromScalar<1> for Peripheral {
 pub struct SingleTransfer {
     pub peripheral: Peripheral,
     pub write_data: Vec<u8>,
-    pub read_len: u8,
+    pub read_len: usize,
 }
